Adding a timeout for SSH and only using terraform destroy if state file exist

+++ tests/devstack/destroy.sh
@@ -1,2 +1,4 @@
 #!/bin/bash
-SSL_CERT_FILE=pebble.minica.pem terraform destroy -auto-approve
+if [ -f "terraform.tfstate" ]; then
+    SSL_CERT_FILE=pebble.minica.pem terraform destroy -auto-approve
+fi
